apisite/user/serialize.py
from rest_framework import serializers
from .models import User
from django.contrib.auth.hashers import make_password

class UserSerializer(serializers.ModelSerializer):
    def create(self, validated_data):
        validated_data['password'] = make_password(validated_data['password'])
        user = User.objects.create(**validated_data)
        return user

    def validate(self, attrs):
        return attrs
    class Meta:
        model = User
        fields = ('email', 'password', 'nickname')