apisite/requirements.txt
asgiref==3.4.1
Django==3.1.3
djangorestframework==3.13.1
mysqlclient==2.1.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
